# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Go Delivery API",
    "description": "Tenant-aware delivery and commerce API with multi-merchant support",
    "version": "1.0.0",
    "contact": {
      "email": "support@godelivery.com"
    }
  },
  "servers": [
    { "url": "http://localhost:3000/api/v1", "description": "Development" },
    { "url": "https://api.godelivery.com/api/v1", "description": "Production" }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "CreateProductDto": {
        "type": "object",
        "required": ["category_id", "name", "base_price"],
        "properties": {
          "category_id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "base_price": { "type": "number", "minimum": 0 },
          "image_url": { "type": "string" },
          "track_inventory": { "type": "boolean", "default": false }
        }
      },
      "ProductResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "merchant_id": { "type": "string", "format": "uuid" },
          "category_id": { "type": "string", "format": "uuid" },
          "category_name": { "type": "string" },
          "name": { "type": "string" },
          "description": { "type": "string", "nullable": true },
          "base_price": { "type": "number" },
          "image_url": { "type": "string", "nullable": true },
          "track_inventory": { "type": "boolean" },
          "is_active": { "type": "boolean" },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
      "CreateAddonDto": {
        "type": "object",
        "required": ["name", "price"],
        "properties": {
          "name": { "type": "string" },
          "price": { "type": "number", "minimum": 0 }
        }
      },
      "AddonResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "product_id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "price": { "type": "number" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
...
      "CreateServiceZoneDto": {
        "type": "object",
        "required": ["zone_id", "branch_id"],
        "properties": {
          "zone_id": { "type": "string", "format": "uuid" },
          "branch_id": { "type": "string", "format": "uuid" }
        }
      },
      "CheckCoverageDto": {
        "type": "object",
        "required": ["latitude", "longitude"],
        "properties": {
          "latitude": { "type": "number" },
          "longitude": { "type": "number" }
        }
      },
...
    }
  },
  "paths": {
...
    "/merchant/products": {
      "get": {
        "tags": ["Products"],
        "summary": "List products",
        "parameters": [{ "name": "x-merchant-id", "in": "header", "schema": { "type": "string", "format": "uuid" } }],
        "responses": {
          "200": { "description": "Success", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/ProductResponse" } } } } }
        }
      },
      "post": {
        "tags": ["Products"],
        "summary": "Create product",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateProductDto" } } }
        },
        "responses": {
          "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProductResponse" } } } }
        }
      }
    },
    "/products/{product_id}": {
      "get": {
        "tags": ["Products"],
        "summary": "Get product details",
        "parameters": [{ "name": "product_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": {
          "200": { "description": "Success", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProductResponse" } } } }
        }
      }
    },
    "/products/{product_id}/addons": {
      "get": {
        "tags": ["Products"],
        "summary": "List product addons",
        "parameters": [{ "name": "product_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": {
          "200": { "description": "Success", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/AddonResponse" } } } } }
        }
      },
      "post": {
        "tags": ["Products"],
        "summary": "Add addon to product",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "product_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateAddonDto" } } }
        },
        "responses": {
          "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AddonResponse" } } } }
        }
      }
    },
...
    "/merchant/service-zones": {
      "get": {
        "tags": ["Service Zones"],
        "summary": "List merchant service zones",
        "parameters": [{ "name": "x-merchant-id", "in": "header", "schema": { "type": "string", "format": "uuid" } }],
        "responses": {
          "200": { "description": "Success" }
        }
      },
      "post": {
        "tags": ["Service Zones"],
        "summary": "Assign zone to branch",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateServiceZoneDto" } } }
        },
        "responses": {
          "201": { "description": "Created" }
        }
      }
    },
    "/merchant/service-zones/check": {
      "post": {
        "tags": ["Service Zones"],
        "summary": "Check delivery coverage",
        "parameters": [{ "name": "x-merchant-id", "in": "header", "schema": { "type": "string", "format": "uuid" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CheckCoverageDto" } } }
        },
        "responses": {
          "200": { "description": "Coverage check result" }
        }
      }
    },
...
  }
}